backend/requirements.txt
fastapi
uvicorn
python-multipart
pdf2image
chromadb
langchain-groq
langchain-huggingface
python-dotenv
docling
sentence-transformers
pydantic 